% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_checks.R
\name{checkArgClass}
\alias{checkArgClass}
\title{check that an object is of the epected class, or stop}
\usage{
checkArgClass(x, expected, msg = "")
}
\arguments{
\item{x}{object to check}

\item{expected}{character, expected class of object x}

\item{msg}{character, message printed at beginning of stop message}
}
\description{
check that an object is of the epected class, or stop
}
\keyword{internal}
