% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_checks.R
\name{checkNotNull}
\alias{checkNotNull}
\title{check that object x is not null, or stop}
\usage{
checkNotNull(x, msg)
}
\arguments{
\item{x}{object to check}

\item{msg}{character, message printed at beginning of stop message}
}
\description{
check that object x is not null, or stop
}
\keyword{internal}
