% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_stats.R
\name{contrast.test}
\alias{contrast.test}
\title{Compute two p-values from a list of numbers emphasizing contrasts}
\usage{
contrast.test(li, contrast = 1)
}
\arguments{
\item{li}{ordered list with numeric values, length must be at least 2}

\item{contrast}{integer, number of middle groups that will be omitted
(when total number of groups is odd, the groups skipped will be contrast+1)}
}
\description{
the first p-value is an ranked anova p
the second p-value is a high-contrast wilcoxon p
for a list with several elements, the wilcoxon groups will be: examples:
(1), 2, (3)
(1), 2, 3, (4)
(1, 2), 3, (4, 5)
(1, 2, 3, 4), 5, 6, (7, 8, 9, 10)
The groupings can be changed by increasing the contrast, e.g. contrast=2
(1), 2, 3, (4)
(1), 2, 3, 4, (5)
In these examples, the groups witin parenthese are merged together, then
the two parentheses are compared using wilcoxon rank sum test
}

