% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.canberra}
\alias{dist.canberra}
\title{Canberra distance}
\usage{
dist.canberra(x)
}
\arguments{
\item{x}{numeric matrix}
}
\description{
This is a wrapper for stats::dist with method="canberra".
The wrapper replaces NA and Inf values with mean and max guesses.
}
