% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.clust}
\alias{dist.clust}
\title{Distance that incorporates weights from a clustering (hclust or pam)}
\usage{
dist.clust(x, clust.dist = "euclidean", clust.method = "complete",
  clust.k = 2, clust.weight = 0.5, clust.alt = FALSE)
}
\arguments{
\item{x}{numeric matrix}

\item{clust.dist}{character, determines what preliminary distance function is
applied prior to clustering}

\item{clust.method}{character, determines agglomeration method in hclust (e.g. complete,
single, average) or "pam"}

\item{clust.k}{integer, determines depth of clustering weights}

\item{clust.weight}{numeric, determines weighting of cluster distances relative to
neighbor rank distance. Set to 0 to obtain pure neighbor rank distance (clustering is ignored).
Set to Inf to obtain pure cluster-based distance. Default is 0.5 to nudge neighbor rank distance
in the direction of the cluster distance.}

\item{clust.alt}{boolean. Set TRUE to obtain an alternative clustering. Default is FALSE
which returns distances in which patterns from a usual distance are reinforced by clustering.}
}
\description{
Distance that incorporates weights from a clustering (hclust or pam)
}
