% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.correlation}
\alias{dist.correlation}
\title{Compute dissimilarity using correlation}
\usage{
dist.correlation(x, directional = TRUE, use.ranks = FALSE)
}
\arguments{
\item{x}{matrix of values}

\item{directional}{logical, set TRUE to get dissimilarities as (1-r) and FALSE
to compute dissimilarities as (1-abs(r))}

\item{use.ranks}{logical, set TRUE to use rho correlation for ranks}
}
\description{
Compute dissimilarity using correlation
}
