% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.dbscan}
\alias{dist.dbscan}
\title{Distance that incorporates weights from a dbscan clustering}
\usage{
dist.dbscan(x, eps = 1, clust.weight = 0.8)
}
\arguments{
\item{x}{numeric matrix}

\item{eps}{numeric, passed to dbscane eps}

\item{clust.weight}{numeric, determines weighting of cluster distances relative to
neighbor rank distance.
Set to 0 to obtain pure neighbor rank distance (clustering is ignored).
Set to Inf to obtain pure cluster-based distance.
Default is 0.5 to nudge neighbor rank distance in the direction of the cluster distance.}
}
\description{
Distance that incorporates weights from a dbscan clustering
}
