% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.hamming}
\alias{dist.hamming}
\title{Hamming distance}
\usage{
dist.hamming(x)
}
\arguments{
\item{x}{matrix with numbers, characters, or factors}
}
\description{
This counts the number of columns in x that are different between rows.
It is a suitable distance function for factor and character data.
}
