% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.log2euclidean}
\alias{dist.log2euclidean}
\title{Euclidean distance of log2 transformed values}
\usage{
dist.log2euclidean(x, shift = 1)
}
\arguments{
\item{x}{numeric matrix}

\item{shift}{numeric, function computes log2(x+shift) before
applying euclidean distance}
}
\description{
Euclidean distance of log2 transformed values
}
