% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.manhattan}
\alias{dist.manhattan}
\title{Manhattan distance}
\usage{
dist.manhattan(x)
}
\arguments{
\item{x}{numeric matrix}
}
\description{
This is a wrapper for stats::dist with method="manhattan".
The wrapper replaces NA and Inf values with mean and max guesses.
}
