% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.pc1}
\alias{dist.pc1}
\title{Distance using 1st principal component from PCA}
\usage{
dist.pc1(x)
}
\arguments{
\item{x}{numeric matrix}
}
\description{
Distance using 1st principal component from PCA
}

