% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.pearson}
\alias{dist.pearson}
\title{Compute distance using pearson correlation r}
\usage{
dist.pearson(x, directional = TRUE)
}
\arguments{
\item{x}{matrix of values}

\item{directional}{logical, set TRUE to get dissimilarities as (1-r) and FALSE
to compute dissimilarities as (1-abs(r))}
}
\description{
Compute distance using pearson correlation r
}
