% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.perm}
\alias{dist.perm}
\title{Random distance obtained through permutation of an input matrix}
\usage{
dist.perm(x, dist.method = c("euclidean", "canberra", "spearman"),
  perm.method = c("shuffle", "bootstrap"))
}
\arguments{
\item{x}{numeric matrix}

\item{dist.method}{character, determines how the distance will be computed,
e.g. euclidean, canberra, etc.}

\item{perm.method}{character, determines how the permutation is carried out,
e.g. free, marginals}
}
\description{
Random distance obtained through permutation of an input matrix
}

