% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.rbeta}
\alias{dist.rbeta}
\title{Random distance using beta distribution}
\usage{
dist.rbeta(x, alpha = 0.5, beta = 0.5)
}
\arguments{
\item{x}{numeric matrix (see stats::dist for details) The object is used to obtain the
size of the random distance object}

\item{alpha}{numeric, shape parameter for beta distribution}

\item{beta}{numeric, shape parameter for beta distribution}
}
\description{
This function generates points in a d-dimensional unit-space with coordinates
chosen at random from a beta distribution. The function then computes the euclidean
distance in this space.
}
