% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.rnorm}
\alias{dist.rnorm}
\title{Random distance}
\usage{
dist.rnorm(x)
}
\arguments{
\item{x}{numeric matrix (see stats::dist for details) The object is used to obtain the
size of the random distance object}
}
\description{
This produces a distances object in which distances are set to random numbers
from the N(0,1) distribution
}

