% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.rnorm0}
\alias{dist.rnorm0}
\title{Random distance using a normal distribution}
\usage{
dist.rnorm0(x, d = round(log2(nrow(x))))
}
\arguments{
\item{x}{numeric matrix (see stats::dist for details) The object is used to obtain the
size of the random distance object}

\item{d}{integer, dimension of the underlying space}
}
\description{
This function generates points in a d-dimensional unit-space with coordinates
chosen at random from a normal distribution. The function then computes the euclidean
distance in this space.
}

