% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{dist.spearman}
\alias{dist.spearman}
\title{Compute distance using spearman rho}
\usage{
dist.spearman(x)
}
\arguments{
\item{x}{matrix of values}
}
\description{
This function rows as features and columns as features.
}
