% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plugins.R
\name{hamming.MultiPatternPlugin}
\alias{hamming.MultiPatternPlugin}
\title{Add a set of configurations to a MultiPattern object}
\usage{
hamming.MultiPatternPlugin(MP, data.name, config.prefix,
  preprocess.prefix = "", preprocess = NULL)
}
\arguments{
\item{MP}{MultiPattern}

\item{data.name}{character}

\item{config.prefix}{character}

\item{preprocess.prefix}{character}

\item{preprocess}{character, (ignored)}
}
\description{
This plugin adds several configurations that use hamming distance on
columns in the data that are factors or characters
}
