% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plugins.R
\name{manhattan.MultiPatternPlugin}
\alias{manhattan.MultiPatternPlugin}
\title{Add a set of configurations to an MultiPattern object}
\usage{
manhattan.MultiPatternPlugin(MP, data.name, config.prefix,
  preprocess.prefix = "", preprocess = NULL)
}
\arguments{
\item{MP}{MultiPattern}

\item{data.name}{character}

\item{config.prefix}{character}

\item{preprocess.prefix}{character}

\item{preprocess}{character}
}
\description{
This plugin adds one configuration that uses a traditional manhattan distance.
}

