% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_enrich.R
\name{MPaddconcepts}
\alias{MPaddconcepts}
\title{Manipulates an MPconcepts list}
\usage{
MPaddconcepts(MPc, item2concept = NULL, concept2item = NULL,
  conceptmatrix = NULL)
}
\arguments{
\item{MPc}{and MPconcepts object}

\item{item2concept}{list with concept membership,
e.g. list(concept1=c(item1, item2), etc.)}

\item{concept2item}{list with concept membership,
e.g. list(item1=c(concept1, concept2), etc.)}

\item{conceptmatrix}{matrix or data frame with membership.
Rows are trated as items, column names as features/concepts.}
}
\description{
Manipulates an MPconcepts list
}

