% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_management.R
\name{MPaddConfig}
\alias{MPaddConfig}
\title{Add one or more cluster configurations into a MultiPattern object}
\usage{
MPaddConfig(MP, config.name, data.name = names(MP$data)[1],
  preprocess = NULL, dist.fun = dist.euclidean)
}
\arguments{
\item{MP}{existing MP object}

\item{config.name}{character, base name for new subspace/distance}

\item{data.name}{character, name for data component in MP object. Defaults to first dataset
declared in MP.}

\item{preprocess}{function or feature set, data preprocessing applied before distance function.
Can be a single value or a list. Defaults to NULL, which indicates no preprocessing is required.
When a function, the function is applied on the dataset prior to computing dissimilarities.
When a vector (character or integer), it is interpreted a feature set of the dataset;
dissimilarities are then computed on this subspace of the data.}

\item{dist.fun}{function, distance function. Can be a single value or a list.
Defaults to euclidean distance.}
}
\description{
The function acceps data.fun or dist.fun as a list. In these cases, the
configuration names are given as config.name:[id].
}

