% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_management.R
\name{MPaddData}
\alias{MPaddData}
\title{Add a dataset into a MultiPattern configuration object}
\usage{
MPaddData(MP, data)
}
\arguments{
\item{MP}{an existing MultiPattern object}

\item{data}{named list with data objects.}
}
\description{
Add a dataset into a MultiPattern configuration object
}
