% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPaddPCAset}
\alias{MPaddPCAset}
\title{Creates a set of configurations using prcomp}
\usage{
MPaddPCAset(MP, dd, ddname)
}
\arguments{
\item{MP}{MultiPattern}

\item{dd}{matrix, dataset in MP to process}

\item{ddname}{character, name of dataset}
}
\description{
Creates a set of configurations using prcomp
}

