% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_management.R
\name{MPchangeSettings}
\alias{MPchangeSettings}
\title{Change settings encoded in a MultiPattern object}
\usage{
MPchangeSettings(MP, settings = list())
}
\arguments{
\item{MP}{a MultiPattern configuration object}

\item{settings}{list with new settings

Accepted settings names are 'num.PCs', 'num.random', 'alpha',
'rpca.term.delta', 'clust.k'}
}
\description{
Change settings encoded in a MultiPattern object
}

