% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPcontrast.test}
\alias{MPcontrast.test}
\title{Compute two p-values from a list of numbers emphasizing contrasts}
\usage{
MPcontrast.test(li)
}
\arguments{
\item{li}{ordered list with}
}
\description{
the first p-value is an ranked anova p
the second p-value is a high-contrast wilcoxon p
for a list with several elements, the wilcoxon groups will be: examples:
(1), 2, 3, (4)
(1, 2), 3, (4, 5)
(1, 2, 3, 4), 5, 6, (7, 8, 9, 10)
}

