% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPcutree}
\alias{MPcutree}
\title{Provides a cut of hclust into k clusters}
\usage{
MPcutree(tree, k = 2, minsize = 1/(10 * k))
}
\arguments{
\item{tree}{tree object output by hclust}

\item{k}{integer, number of clusters to output}

\item{minsize}{numeric in range [0, 1], minimal size of the clusters}
}
\description{
In contrast to stats::cutree(), this can output more than k clusters. However, the
output should contain k clusters that are of nontrivial size. The largest cluster
should be labeled as "1", the next biggest as "2" and so on
}

