% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_datasets.R
\docType{data}
\name{MPdata4L}
\alias{MPdata4L}
\title{Toy dataset with four groups (large)}
\format{A data frame with three columns: \code{D1}, \code{D2}, \code{class}.}
\usage{
data(MPdata4L)
}
\description{
A synthetic collection of points in 2D with expected cluster groups.
}
\keyword{datasets}
