% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_datasets.R
\docType{data}
\name{MPdata6S}
\alias{MPdata6S}
\title{Toy dataset with six groups (small)}
\format{A data frame with three columns: \code{D1}, \code{D2}, \code{class}.}
\usage{
data(MPdata6S)
}
\description{
A synthetic collection of points in 2D with expected cluster groups.
}
\keyword{datasets}

