% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_management.R
\docType{data}
\name{MPdefaultSettings}
\alias{MPdefaultSettings}
\title{Summary of default settings for new MultiPattern objects}
\format{An object of class \code{list} of length 12.}
\usage{
MPdefaultSettings
}
\description{
num.random - number of random configurations
}
\details{
num.PCs - number of PCA components

rpca.term.delta - used in rpca analysis (smaller than default for speed)

clust.k - max number of clusters to use in easyConfig reg and alt
configurations

nmf.bg - background level in nmf analysis (avoids all-zero rows)

nmf.rank - maximal rank of NMF analysis (zero triggers an
automatic decision)

subspace.num.random - number of random subspaces

subspace.d.random - number/proportion of features for subspace analysis

alpha - exponent for similarity transformation for meta-similarities

beta - determines Lbeta distance for meta-similarities

subsample.N - number of observations to use in subsample/bootstrap

subsample.R - number of subsampling/bootstrap repetitions
}
\keyword{datasets}

