% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{MPdistImpute}
\alias{MPdistImpute}
\title{Impute values for NAs in a dist object}
\usage{
MPdistImpute(x, method = mean)
}
\arguments{
\item{x}{matrix or dist object}

\item{method}{function, is applied on x to find the imputed values. This
function must be able to perform method(x, na.rm=TRUE)}
}
\description{
(This is a specialized function - it converts x into a dist as the first step.
This is to enforce consistency when applying the 'method' function.
e.g. to avoid confusion between mean(matrix) and mean(as.dist(matrix)), which are not the same)
}
