% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_enrich.R
\name{MPenrichment}
\alias{MPenrichment}
\title{Computes simple Fisher-based enrichment calculations using a set of hits, a set of
background items, and concept definitions}
\usage{
MPenrichment(MPconcepts, hits, background)
}
\arguments{
\item{MPconcepts}{list linking items to concepts}

\item{hits}{character vector with items}

\item{background}{character vector with items (background set)}
}
\description{
Computes simple Fisher-based enrichment calculations using a set of hits, a set of
background items, and concept definitions
}

