% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPgetAll2Subspaces}
\alias{MPgetAll2Subspaces}
\title{obtain a list of all pairs from a set}
\usage{
MPgetAll2Subspaces(f)
}
\arguments{
\item{f}{vector of features (or numbers for column names)}
}
\description{
obtain a list of all pairs from a set
}

