% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_management.R
\name{MPgetDistances}
\alias{MPgetDistances}
\title{Run MultiPattern analysis.}
\usage{
MPgetDistances(MP, verbose = TRUE)
}
\arguments{
\item{MP}{MultiPattern configuration object}

\item{verbose}{logical, when TRUE function prints a warning message when the input is large}
}
\description{
This function runs analyses specified in a MultiPattern configuration object.
It returns an object holding similarity matrices for each of the MultiPattern analyses.
}

