% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPgetKNearest}
\alias{MPgetKNearest}
\title{Extract the k items that are nearest to specified seeds}
\usage{
MPgetKNearest(nowdist, seeds, k)
}
\arguments{
\item{nowdist}{a distance object or matrix}

\item{seeds}{names of seed items}

\item{k}{number of nearest neighbors to select}
}
\description{
Extract the k items that are nearest to specified seeds
}
