% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPgetMap}
\alias{MPgetMap}
\title{Create a map for a distance object/matrix}
\usage{
MPgetMap(d, method = c("umap", "cmdscale"), ...)
}
\arguments{
\item{d}{distance object}

\item{method}{character, method to prepare map}

\item{...}{other parameters can be passed to umap}
}
\description{
This function uses cmdscale, or tsne.
}
