% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_management.R
\name{MPgetMetaDistance}
\alias{MPgetMetaDistance}
\title{Compare MultiPattern similarity matrices}
\usage{
MPgetMetaDistance(MPS, standardize = MPrankNeighbors, alpha = 1, beta = 2,
  ...)
}
\arguments{
\item{MPS}{a MultiPatternSimilarities object (list of similarity matrices)}

\item{standardize}{function used to transform similarity matrices}

\item{alpha}{numeric, used as an alternative way to transform similarity matrices,
applied after transform.fun}

\item{beta}{numeric, sets the Lbeta norm}

\item{...}{additional arguments, used when applying standardize function to the MPS}
}
\description{
Compare MultiPattern similarity matrices
}

