% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_neighbors.R
\name{MPgetNeighborSet}
\alias{MPgetNeighborSet}
\title{Get a set of nearest neighbors (ranked)}
\usage{
MPgetNeighborSet(dd, seedsample, maxrank = 0, ...)
}
\arguments{
\item{dd}{- distance object}

\item{seedsample}{- name of sample}

\item{maxrank}{- integer/numeric. Leave 0 to obtain complete ranking. Set to N if
interested only in items with rank <=N}

\item{...}{- passed on to rank()}
}
\value{
vector with names, values are ranks, names correspond to elements in dd
}
\description{
This function scores graph similarity using rank neighborhoods
}

