% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPgetNonredundantGenesets}
\alias{MPgetNonredundantGenesets}
\title{get a subset of genesets that are small size and non-redundant}
\usage{
MPgetNonredundantGenesets(sets, target.genes = NULL, target.size = c(2, 20),
  target.JI = 0.2)
}
\arguments{
\item{sets}{list with gene sets}

\item{target.genes}{character vector, names of ok gene names}

\item{target.size}{integer vector of length 2, indicate minimum and maximum size of geneset}

\item{target.JI}{numeric, maximum Jaccard overlap between the new genesets}
}
\description{
This is an inelegant implementation, maybe fix and simplify later.
}

