% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPgetRandomSubspaces}
\alias{MPgetRandomSubspaces}
\title{Produce a list of randomly selected feature sets}
\usage{
MPgetRandomSubspaces(f, n, d = 4, oversample = 4)
}
\arguments{
\item{f}{character or integer vector, feature set}

\item{n}{integer, number of required feature sets}

\item{d}{integer, dimension of the required subspaces}

\item{oversample}{numeric, used in randomization process (only becomes relevant
when function starts to return fewer subspaces than requested, see description above)}
}
\description{
Note: this function is not guaranteed to produce the specified
number of features (it might produce less).
}
\details{
For example, if a feature set is of length 3 and user requets 5 sets of
size 3, the output will contain only 1 (because there are not others).

If a feature set is of length 6 and user requests 20 sets of length 3,
the output might contain fewer than 20. Although in this case there are exactly 20
sets of length 3, the randomization might not find them all. In such cases, try
increasing the setting oversample.
}
