% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_management.R
\name{MPgetRepresentatives}
\alias{MPgetRepresentatives}
\title{Identify representative points from a distance object}
\usage{
MPgetRepresentatives(d, subsets = NULL, method = c("complete", "single",
  "average", "pam", "extreme"), k = 0.05)
}
\arguments{
\item{d}{dist object}

\item{subsets}{if NULL, function considers all rows
if vector of characters, interpreted as regex for subsetting rows}

\item{method}{character, determines what clustering approach
is used to determine representative items within a clustering}

\item{k}{numeric of integer. If more than 1, number of points to select
If less than 1, proportion of points to select}
}
\description{
Identify representative points from a distance object
}

