% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPgetTopPC}
\alias{MPgetTopPC}
\title{Get information about top features contributing to PC components}
\usage{
MPgetTopPC(x, pc = 1, n = NULL, plot = FALSE, mar = c(4, 24, 4, 4))
}
\arguments{
\item{x}{prcomp object}

\item{pc}{integer, the principal component to "explain", can be a vector of integers}

\item{n}{integer, number of features to output}

\item{plot}{logical, when true it automatically plots the data}

\item{mar}{margin values}
}
\description{
Get information about top features contributing to PC components
}

