% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_distances.R
\name{MPimputeNAs}
\alias{MPimputeNAs}
\title{Impute values for NAs in a matrix or other object}
\usage{
MPimputeNAs(x, method = mean)
}
\arguments{
\item{x}{matrix or dist object}

\item{method}{function, is applied on x to find the imputed values. This
function must be able to perform method(x, na.rm=TRUE)}
}
\description{
(This is a general-purpose function and works on many types of objects with some NAs.
The object must support operations: sum, method (e.g. mean), and subsetting via []).
}

