% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPingenesets}
\alias{MPingenesets}
\title{shows the genesets that contain genes}
\usage{
MPingenesets(sets, genes)
}
\arguments{
\item{sets}{list of gene sets}

\item{genes}{vector of genes, all genes must be present in a set to provide a hit}
}
\description{
shows the genesets that contain genes
}

