% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plugins.R
\name{MPlistPlugins}
\alias{MPlistPlugins}
\title{Get of MultiPattern plugins available}
\usage{
MPlistPlugins()
}
\description{
This returns a vector of strings. A string x indicates that there exist
an accessible function called x.MultiPatternPlugin.
}

