% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPmakePrepFunction}
\alias{MPmakePrepFunction}
\title{Create a function for preprocessing a dataset before computing a similarity}
\usage{
MPmakePrepFunction(a)
}
\arguments{
\item{a}{various types of inputs are accepted.
If a is a function, output will be exactly a.
If a is a character vector, output is a function that returns the a-columns in a matrix.
If a is an integer vector}
}
\description{
Create a function for preprocessing a dataset before computing a similarity
}

