% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPmatrixColNorm}
\alias{MPmatrixColNorm}
\title{Normalize a matrix column-wise by median shift and division by sd}
\usage{
MPmatrixColNorm(x)
}
\arguments{
\item{x}{numeric matrix or data.frame}
}
\description{
Normalize a matrix column-wise by median shift and division by sd
}
