% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_management.R
\name{MPnew}
\alias{MPnew}
\title{Create a basic MP object}
\usage{
MPnew(items, data = NULL)
}
\arguments{
\item{items}{character vector, specifies names for observations in the multipatter analysis}

\item{data}{named list with data objects. Specifying this parameter
is equivalent to calling MPnew() and then adding data manually with MPaddData()}
}
\description{
Create a basic MP object
}

