% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_suggest.R
\name{MPPartitionFeatures}
\alias{MPPartitionFeatures}
\title{Partition features in a data matrix into categories like "real",
"realskew", "bin", "multi", etc.}
\usage{
MPPartitionFeatures(data)
}
\arguments{
\item{data}{matrix of data}
}
\value{
named list, names indicate feature class, contents are features
}
\description{
Partition features in a data matrix into categories like "real",
"realskew", "bin", "multi", etc.
}
