% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plot.R
\name{MPplotmap}
\alias{MPplotmap}
\title{Custom function to draw mds maps. Uses Rcssplot and some hard settings}
\usage{
MPplotmap(xylayout, color = c(), highlight.points = c(), xypadding = 0.05,
  main = "", legend.separate = FALSE, squarexy = TRUE, label = FALSE,
  Rcss = "default", Rcssclass = c(), ...)
}
\arguments{
\item{xylayout}{matrix or list. When matrix, should contain two columns x/y for positions of
points on 2d map. When list, should contain several matrices as just described.}

\item{color}{list with names. Should contain vectors of items in xylayout.
These items will be plotted with the css style determined by names(color)}

\item{highlight.points}{vector of rows in the xylayout that will be
plotted with the `highlight` style}

\item{xypadding}{numeric, determines empty space between points and box}

\item{main}{character, title for the plot}

\item{legend.separate}{logical, set TRUE to create a two-panel display. Highlighted
configurations will be labeled as A, B, C in main panel, then explained in the second panel.}

\item{squarexy}{logical, set TRUE to force the scales on x and y axis to be the same
(recommended so that euclidean distances on plot correspond to euclidean distances in
layout algos)}

\item{label}{logical, TRUE to print names of points}

\item{Rcss}{Rcss object}

\item{Rcssclass}{Rcss class vector}

\item{...}{arguments passed on to plot (e.g. xlab, ylab)}
}
\description{
Custom function to draw mds maps. Uses Rcssplot and some hard settings
}
