% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plot.R
\name{MPplotPairwiseSets}
\alias{MPplotPairwiseSets}
\title{plot pairwise overlaps summaries}
\usage{
MPplotPairwiseSets(ds, samplenames, seedsample, neighborhood = 10,
  ds.names = NULL, bg = NULL, plot = TRUE, order = TRUE,
  density.n = 128, density.adjust = 2, Rcss = "default",
  Rcssclass = c())
}
\arguments{
\item{ds}{list of distance matrices}

\item{samplenames}{vector with samples contained in distance objects}

\item{seedsample}{one or more sample name (produces one plot for each seed)}

\item{neighborhood}{integer. Size of neighborhood around seed sample}

\item{ds.names}{vector with labels to replace ds original names}

\item{bg}{vector with labels from ds.names that are treated as class 'bg'}

\item{plot}{logical. Set TRUE to get figures. Set FALSE to get comparison data frame.}

\item{order}{logical. Set TRUE to arrange multiple seedsample by sum JI}

\item{density.n}{integer. Passed on to density to compute smoothing fo JI when seedsample=NA}

\item{density.adjust}{integer}

\item{Rcss}{Rcss object for Rcssplot}

\item{Rcssclass}{class for Rcssplot}
}
\description{
e.g. set A overlaps with B by 50%, with C by 30%, with D by 70%, etc.
}

