% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plot.R
\name{MPplotScatterWithK}
\alias{MPplotScatterWithK}
\title{plot a 2D scatter diagram with clusters in various style}
\usage{
MPplotScatterWithK(coords, clust, Gnames = TRUE, main = "",
  RC = "default", RCC = c())
}
\arguments{
\item{coords}{matrix, first two columns will be interpreted as xy coordinates}

\item{clust}{character, mapping between items and cluster numbers}

\item{Gnames}{logical, set TRUE to force cluster group names into G1, G2, etc.
set FALSE to use the existing group names in clust}

\item{main}{character, displayed as plot title}

\item{RC}{Rcss object}

\item{RCC}{Rcss class}
}
\description{
plot a 2D scatter diagram with clusters in various style
}

