% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plot.R
\name{MPplotScatterWithLinks}
\alias{MPplotScatterWithLinks}
\title{plot a 2D scatter diagram with clusters in various style}
\usage{
MPplotScatterWithLinks(coords, xyd, seed = rownames(coords)[1],
  seed.links = round(nrow(coords)/4), main = "", Rcss = "default",
  Rcssclass = c())
}
\arguments{
\item{coords}{matrix, first two columns will be interpreted as xy coordinates}

\item{xyd}{distance object or matrix}

\item{seed}{character, sample for which to show neighbors}

\item{seed.links}{integer, number of links}

\item{main}{character, displayed as plot title}

\item{Rcss}{Rcss object}

\item{Rcssclass}{Rcss class}
}
\description{
plot a 2D scatter diagram with clusters in various style
}

