% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_neighbors.R
\name{MPrankNeighbors}
\alias{MPrankNeighbors}
\title{Compute non-parametric neighborhoods of data elements using ranks on a per column basis}
\usage{
MPrankNeighbors(dd, ties = "average", ...)
}
\arguments{
\item{dd}{- matrix of distances between neighbors}

\item{ties}{- character, used to determine how equal ranks are handled by rank function}

\item{...}{- additional arguments passed on to rank}
}
\description{
This function takes as input a distance object or distance matrix (values
describing distance between pairs of data elements). For each data element,
it considers what other points are nearby and ranks them by distance.
}
\details{
The output is non-parametric neighborhood object. "x[i,j] = r" implies that
data element "j" is the "r-th" one to data element "i".
}
