% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_neighbors.R
\name{MPrankWhole}
\alias{MPrankWhole}
\title{Transform a dissimilariy matrix using ranks}
\usage{
MPrankWhole(dd, ties.method = "min", ...)
}
\arguments{
\item{dd}{matrix of distances between neighbors}

\item{ties.method}{character, used to determine how equal ranks are handled by rank function.}

\item{...}{additional arguments passed on to rank}
}
\description{
This function takes as input a distance object or distance matrix (values
describing distance between pairs of data elements). For each data element,
it considers what other points are nearby and ranks them by distance.
}

