% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_management.R
\name{MPremove}
\alias{MPremove}
\title{Remove data or configuration components from a MultiPattern object}
\usage{
MPremove(MP, data = NULL, config = NULL)
}
\arguments{
\item{MP}{a MultiPattern configuration objects}

\item{data}{character or vector of data identifiers}

\item{config}{character or vector of config identifiers}
}
\description{
Remove data or configuration components from a MultiPattern object
}
