% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPsquarelim}
\alias{MPsquarelim}
\title{Create equal size xlim and ylim intervals from x,y values}
\usage{
MPsquarelim(x, y)
}
\arguments{
\item{x}{values for the x axis}

\item{y}{values for the y axis}
}
\description{
This function is useful for creating MDS diagrams in a square format,
which should have x and y axes in the same scale.
}
