% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_suggest.R
\name{MPsuggestConfigLong}
\alias{MPsuggestConfigLong}
\title{Suggest a set of clustering analyses for a dataset}
\usage{
MPsuggestConfigLong(MP, data, random.perm = 0, random.generic = 100,
  clust.types = c("average", "pam"))
}
\arguments{
\item{MP}{MultiMetric configuration object}

\item{data}{character. Name of dataset defined in MP.}

\item{random.perm}{integer. Set to number of randomized configurations for each
suggested analysis type}

\item{random.generic}{integer. Number of generic randomized configurations}

\item{clust.types}{character. Types of clustering-based distances to 

I think this is deprecated}
}
\description{
This function can take some time to execute for large datasets. This is
because the function applies principal component decompositions to parts of the
input data and stores the intermediate results in the MP object. This computation
speeds up subsequent steps of the MultiMetric analysis.
}

