% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPsuggestTopFeatures}
\alias{MPsuggestTopFeatures}
\title{Make a suggestion for interesting features in a dataset based on PCA weights}
\usage{
MPsuggestTopFeatures(x, ncomp = function(y) {     1 + sqrt(y) },
  iqrfactor = function(y) {     log(y) })
}
\arguments{
\item{x}{dataset or prcomp object}

\item{ncomp}{function, integer, or numeric.
If integer, number of principal components to consider.
If number less than 1, fraction of components to consider.
If function, a transformation on the number of features that returns an integer.}

\item{iqrfactor}{numeric. threshold to consider a feature in a PC as worthy of selection}
}
\description{
This function looks at the features that make up individual PCA components. It
returns those features that contribute more than others to these componets
(according to the weight of the feature compared to the median and IQR interval)
}
