% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plot.R
\name{MPvalMat2ColMat}
\alias{MPvalMat2ColMat}
\title{Convert a matrix of values (range [-Inf,Inf] to colors using transparency)}
\usage{
MPvalMat2ColMat(x, col = c("#0000ff", "#ff0000"), maxval = 1)
}
\arguments{
\item{x}{numeric matrix.}

\item{col}{vector of two colors in #XXXXXX format. First element determines
color associated with negative values. Second element determines color associated
with positive values}

\item{maxval}{numeric. Value for which color reaches saturation}
}
\description{
This funciton is a copyt of E3Val2Col from package ExpCube (Tomasz Konopka)
}
