% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plugins.R
\name{nmf.MultiPatternPlugin}
\alias{nmf.MultiPatternPlugin}
\title{Add a set of configurations to a MultiPattern object}
\usage{
nmf.MultiPatternPlugin(MP, data.name, config.prefix,
  preprocess.prefix = "", preprocess = NULL)
}
\arguments{
\item{MP}{MultiPattern}

\item{data.name}{character}

\item{config.prefix}{character}

\item{preprocess.prefix}{character}

\item{preprocess}{character}
}
\description{
This plugin adds several configurations that use non-negative matrix factorization.	
This function will not change MP if the dataset is not suitable	
for NMF, for example if the dataset contains missing values	or too-few columns.
}
