% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plugins.R
\name{pam.MultiPatternPlugin}
\alias{pam.MultiPatternPlugin}
\title{Add a set of configurations to an MultiPattern object}
\usage{
pam.MultiPatternPlugin(MP, data.name, config.prefix, preprocess.prefix = "",
  preprocess = NULL)
}
\arguments{
\item{MP}{MultiPattern}

\item{data.name}{character}

\item{config.prefix}{character}

\item{preprocess.prefix}{character}

\item{preprocess}{character}
}
\description{
This plugin adds several configurations that use clustering-driven dissimilarities
(using pam clustering). Some configurations (marked reg) create dissimilarities
that reinforce traditional clustering. Other configuration (marked alt) create
dissimilarities that promote alternative clustering. Settings for this plugin
are extracted from MP$settings.
}
