% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plugins.R
\name{pca.ica.helper}
\alias{pca.ica.helper}
\title{helper function to fastica and pca plugins; first few arguments are same
as for MultiPatternPlugin functions. Last two}
\usage{
pca.ica.helper(MP, data.name, config.prefix, preprocess.prefix = "",
  preprocess = NULL, data = NULL, type = c("pca", "ica"))
}
\arguments{
\item{MP}{MultiPattern}

\item{data.name}{character}

\item{config.prefix}{character}

\item{preprocess.prefix}{character}

\item{preprocess}{character}

\item{data}{matrix, data}

\item{type}{character, one of pca or ica}
}
\description{
helper function to fastica and pca plugins; first few arguments are same
as for MultiPatternPlugin functions. Last two
}
