% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plugins.R
\name{pca.MultiPatternPlugin}
\alias{pca.MultiPatternPlugin}
\title{Add a set of configurations to a MultiPattern object}
\usage{
pca.MultiPatternPlugin(MP, data.name, config.prefix,
  preprocess.prefix = "", preprocess = NULL)
}
\arguments{
\item{MP}{MultiPattern}

\item{data.name}{character}

\item{config.prefix}{character}

\item{preprocess.prefix}{character}

\item{preprocess}{character}
}
\description{
This plugin transforms an input dataset using PCA, then adds configurations
that use the transformed data. The configurations will of two types:
individual PC components, e.g. only PC1 or only PC2;
pairs of PC components, e.g. PC1, PC2 (PC1.PC2) or PC2, PC4 (PC2.PC4)
}
\details{
The maximal PC component for the configurations is determined from MP$settings.
}
