% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_stats.R
\name{print.contrast}
\alias{print.contrast}
\title{pretty print of information from a contrast test}
\usage{
\method{print}{contrast}(x, ...)
}
\arguments{
\item{x}{object of class contrast}

\item{...}{additional parameters (ignored)}
}
\description{
pretty print of information from a contrast test
}

