% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_print.R
\name{print.MultiPattern}
\alias{print.MultiPattern}
\title{Print a summary of a MP configuration object}
\usage{
\method{print}{MultiPattern}(x, ...)
}
\arguments{
\item{x}{a MultiPattern configuration object}

\item{...}{additional arguments, not used}
}
\description{
Print a summary of a MP configuration object
}

