% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_print.R
\name{print.MultiPatternSettings}
\alias{print.MultiPatternSettings}
\title{Print a summary of a MP distances objects}
\usage{
\method{print}{MultiPatternSettings}(x, ...)
}
\arguments{
\item{x}{a MultiPatternSettings object (list)}

\item{...}{additional arguments, not used}
}
\description{
Print a summary of a MP distances objects
}
