% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_print.R
\name{print.MultiPatternSimilarities}
\alias{print.MultiPatternSimilarities}
\title{Print a summary of a MP distances objects}
\usage{
\method{print}{MultiPatternSimilarities}(x, ...)
}
\arguments{
\item{x}{a MultiPatternSimilarities object}

\item{...}{additional arguments, not used}
}
\description{
Print a summary of a MP distances objects
}
