% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_management.R
\name{MPchangeSettings}
\alias{MPchangeSettings}
\title{Change settings encoded in a MultiPattern object}
\usage{
MPchangeSettings(MP, settings = list(), warn = TRUE)
}
\arguments{
\item{MP}{a MultiPattern configuration object}

\item{settings}{list with new settings}

\item{warn}{logical, set TRUE to get warnings if a setting value
is not part of the core MultiPattern set (MPdefaultSettings)

Accepted settings names are those defined in MPdefaultSettings.}
}
\description{
Change settings encoded in a MultiPattern object
}
