% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_management.R
\name{MPeasyConfig}
\alias{MPeasyConfig}
\title{Add predefined configuration types to a MultiPattern configuration object}
\usage{
MPeasyConfig(MP, data = NULL, config.prefix = "", preprocess.prefix = "",
  type = c("pca", "euclidean", "spearman", "canberra", "manhattan", "hclust",
  "pam"), random = TRUE, preprocess = NULL)
}
\arguments{
\item{MP}{MultiPattern configuration object}

\item{data}{character, names of datasets in MP to use to create suggestions. If NULL, function
applies the configurations to all datasets (but see type). If not NULL, function only considers
the specified data objects.}

\item{config.prefix}{character, prefix used in all configuration names}

\item{preprocess.prefix}{character, a middle-fix used when naming subspaceR configurations}

\item{type}{character or list, codes for what types of configuration plugins to use. To see all
available plugins, use MPlistPlugins(). 
shows all the supported configuration types. Detailed descriptions will appear elsewhere.}

\item{random}{logical, set TRUE to include random configurations.}

\item{preprocess}{object specifying preprocessing, e.g. vector of features for subspaces}
}
\description{
This function modifies the object MP in the first argument. The primary modifications
are new items in the MP$config list. When pca or rpca are specified in the type argument,
the function also precomputes these transformations and adds these into the MP$data list. 
All configurations relying on pca or rpca transformed data refer to these pre-computed objects.
}
