% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_compute.R
\name{MPgetDistances}
\alias{MPgetDistances}
\title{Run MultiPattern analysis.}
\usage{
MPgetDistances(MP, configs = NULL, verbose = TRUE)
}
\arguments{
\item{MP}{MultiPattern configuration object}

\item{configs}{string, vector of configuration names to evaluate;
leave NULL to evaluate all configurations}

\item{verbose}{logical, when TRUE function prints a warning message when the input is large}
}
\description{
This function runs analyses specified in a MultiPattern configuration object.
It returns an object holding similarity matrices for each of the MultiPattern analyses.
}
