% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPgetMap}
\alias{MPgetMap}
\title{Create a map for a distance object/matrix}
\usage{
MPgetMap(d, whiten = 0.01)
}
\arguments{
\item{d}{distance object}

\item{whiten}{numeric, use low number like 0.01 (default) to slightly jiggle points
on the map. This is a percentage of the x-y ranges.}
}
\description{
This function uses cmdscale, but additionally adds a certain amount of
whitening determined by whiten. The whitening is pseudo-random and deterministic;
it always gives the same result.
}

