% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_helpers.R
\name{MPgetMap}
\alias{MPgetMap}
\title{Create a map for a distance object/matrix}
\usage{
MPgetMap(d, tsne = FALSE, perplexity = 40, max_iter = 1, whiten = FALSE)
}
\arguments{
\item{d}{distance object}

\item{tsne}{logical, determine if use tsne for map layout}

\item{perplexity}{integer, passed on to tsne(), slightly larger than default}

\item{max_iter}{integer, passed on to tsne(), much lower default than in
tsne's default}

\item{whiten}{logical, passed on to tsne(), different default than tsne()}
}
\description{
This function uses cmdscale, or tsne.
}
