% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MP_plugins.R
\name{nmf.MultiPatternPlugin}
\alias{nmf.MultiPatternPlugin}
\title{Add a set of configurations to a MultiPattern object}
\usage{
nmf.MultiPatternPlugin(MP, data.name, config.prefix, preprocess.prefix = "",
  preprocess = NULL)
}
\arguments{
\item{MP}{MultiPattern}

\item{data.name}{character}

\item{config.prefix}{character}

\item{preprocess.prefix}{character}

\item{preprocess}{character}
}
\description{
This plugin transforms an input dataset using rpca, then defines configurations
on the transformed data. The plugin does not do anything if the input data
contains non-finite elements. Settings for the plugin are extracted from MP$settings.
}
\details{
Note: this function will not change MP if the dataset is not suitable
for NMF. This includes situations when the dataset contains missing values
or contains too-few columns.
}
